requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
aws-cdk-lib==2.237.1
aws-cdk.asset-awscli-v1==2.2.263
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==48.20.0
beautifulsoup4==4.14.3
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colorama==0.4.6
constructs==10.4.5
cryptography==46.0.4
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.3
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jsii==1.126.0
mangum==0.21.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
publication==0.0.3
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
ruff==0.15.0
sentry-sdk==2.52.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
typeguard==2.13.3
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
